Revert "Cleanup"
