+++ Cymote/Cymote/Cymote/src/main_sam_b.c
@@ -53,12 +53,16 @@
 #include <asf.h>
 #include "BLE.h"
 #include "LSM9DS0.h"
+#include "BitBang.h"
 #include "ADC_PWM.h"
 #include "timer_hw.h"
 #include "button.h"
 #include "console_serial.h"
 
+
+
 #define USE_BLE 1
+
 
 /* Globals */
 cymote_service_handler_t cymote_service_handler;
@@ -69,6 +73,8 @@ uint8_t dummy_data_counter_old = 0;
 uint8_t dummy_data_2_state = 0;
 uint8_t dummy_data[] = {'A', 'B', 'C', 'D', 'E', 'a'};
 uint16_t accelerometer_data[3];
+
+
 
 /* timer callback function */
 static void timer_callback_fn(void)
@@ -147,22 +153,22 @@ int main(void)
 	acquire_sleep_lock();
 	system_clock_config(CLOCK_RESOURCE_XO_26_MHZ, CLOCK_FREQ_26_MHZ);
 
-	/* Initialize serial console. */
+	/* Initialize serial console */
 	serial_console_init();
 	//configure_console(&console_instance);
 	
-	/* Hardware timer. */
+	/* Hardware timer */
 	hw_timer_init();
 	hw_timer_register_callback(timer_callback_fn);
 	//hw_timer_start(BLE_UPDATE_INTERVAL);
 	
-	/* Button initialization. */
+	/* button initialization */
 	gpio_init();
 	button_init();
 	button_register_callback(button_cb);
 
-	/* LSM9DS0 gpio configuration. */
-	configure_gpio();
+	/* Bit banged gpio configuration. */
+	bb_configure_gpio();
 
 	/* Start up SPI and read the status register. */
 	uint8_t receive = 0x00;
@@ -176,48 +182,51 @@ int main(void)
 	bb_init_accelerometer_scale(A_SCALE_2G);
 	
 	/* Initialize magnetometer control registers. */
-	init_magnetometer();
-	init_magnetometer_odr(M_ODR_100);
-	init_magnetometer_scale(M_SCALE_2GS);
+	bb_init_magnetometer();
+	bb_init_magnetometer_odr(M_ODR_100);
+	bb_init_magnetometer_scale(M_SCALE_2GS);
 	
 	/* Initialize gyroscope control registers. */
-	init_gyroscope();
-	init_gyroscope_odr(G_ODR_190_BW_125);
-	init_gyroscope_scale(G_SCALE_2000DPS);
+	bb_init_gyroscope();
+	bb_init_gyroscope_odr(G_ODR_190_BW_125);
+	bb_init_gyroscope_scale(G_SCALE_2000DPS);
 	
 	
 	if(USE_BLE){
 		DBG_LOG("Initializing BLE Application");
 	
-		/* Initialize the BLE chip and set the device address. */
+		/* initialize the BLE chip  and Set the Device Address */
 		at_ble_addr_t addr;
 		addr.type = AT_BLE_ADDRESS_PUBLIC;
 		int i;
-		for(i = 0; i < AT_BLE_ADDR_LEN; i++){
+		for(i=0; i<AT_BLE_ADDR_LEN; i++){
 			addr.addr[i] = 0;
 		}
-		addr.addr[0] = 1;
+		addr.addr[0]=1;
 		ble_device_init(&addr);
 
 		cymote_init_service(&cymote_service_handler);
-		if((status = cymote_primary_service_define(&cymote_service_handler)) != AT_BLE_SUCCESS){
+		if ((status = cymote_primary_service_define(&cymote_service_handler)) != AT_BLE_SUCCESS)
+		{
 			DBG_LOG("Device Information Service definition failed,reason %x",status);
 		}
 
 		device_information_advertise();
 
-		/* Register callbacks for gap related events. */
-		ble_mgr_events_callback_handler(REGISTER_CALL_BACK, BLE_GAP_EVENT_TYPE, cymote_app_gap_cb);
+		/* Register callbacks for gap related events */
+		ble_mgr_events_callback_handler(REGISTER_CALL_BACK,
+		BLE_GAP_EVENT_TYPE,
+		cymote_app_gap_cb);
 	}
 
-	/* Initialize Joystick registers. */ 
+	/* Initialize Joystick registers */ 
 	configure_adc_pin3();
 	configure_adc_pin4();
 
-    /* Main working loop. */
+    /* Main working loop */
     while (1) 
     {
-		if(USE_BLE){
+		if (USE_BLE){
 			/* BLE Event task */
 			ble_event_task(BLE_EVENT_TIMEOUT);
 		
@@ -229,12 +238,11 @@ int main(void)
 				UPDATE_DUMMY_DATA(&cymote_service_handler, &newData, cymote_connection_handle);
 				dummy_data_counter_old = dummy_data_counter;
 			}
-			
 			if (dummy_data_2_state != 0){
 				dummy_data_2_state = 0;
 			
 				char newValue[DUMMY_DATA_2_MAX_LEN];
-				int value = rand() % 1000;
+				int value = rand()%1000;
 				uint8_t len = snprintf(newValue, DUMMY_DATA_2_MAX_LEN, "%d", value);
 				printf("newValue: %s\r\n", newValue);
 			
@@ -243,13 +251,11 @@ int main(void)
 				UPDATE_DUMMY_DATA_2(&cymote_service_handler, &newData, cymote_connection_handle);
 			}
 		}
-		
 		/*
 		do {
 		} while (adc_read(ADC_INPUT_CH_GPIO_MS1, &result) == STATUS_BUSY);
 		do {
 		} while (adc_read(ADC_INPUT_CH_GPIO_MS2, &result2) == STATUS_BUSY);
-		
 		configure_adc_pin3();
 		configure_adc_pin4();
 		//adc_read(ADC_INPUT_CH_GPIO_MS1, &result);
@@ -265,6 +271,7 @@ int main(void)
 		else
 		configure_pwm_from_duty_pin_11(result2 / 11);
 		*/
+		
 
 		/* Print accelerometer data */
 		bb_print_raw_accelerometer(accelerometer_data);
@@ -295,16 +302,6 @@ int main(void)
 
     }
 }
-
-
-
-
-
-
-
-
-
-
 
 //int main(void)
 //{
